backend/pyproject.toml
[tool.poetry]
name = "gatosnap"
version = "0.1.0"
description = ""
authors = ["Braden Williams <braden@bradenwill.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
plexapi = "^4.14.0"
alembic = "^1.11.1"
pyzmq = "^25.1.0"
greenlet = "^2.0.2"
aiosqlite = "^0.19.0"
fastapi = "^0.100.1"
pydantic = "^2.1.1"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
